动画类由double改为float 枚举类ImageSizeMode部分改为更加合理的名称  其他优化

+++ include/EzUI/Animation.h
@@ -7,18 +7,18 @@ namespace ezui {
 	class UI_EXPORT Animation : public Object {
 	private:
 		Timer *m_timer;
-		double m_startValue = 0;
-		double m_endValue = 0;
-		double m_currValue = 0;
-		double m_speedPerMs = 0;
+		float m_startValue = 0;
+		float m_endValue = 0;
+		float m_currValue = 0;
+		float m_speedPerMs = 0;
 		ULONGLONG m_lastTime = 0;
 	public:
 		//当值更改的时候发生的事件(请绑定此函数进行回调,请自行线程同步)
-		std::function<void(double)> ValueChanged;
+		std::function<void(float)> ValueChanged;
 		Animation(Object* parentObject = NULL);
 		virtual ~Animation();
-		void SetStartValue(double value);
-		void SetEndValue(double value);
+		void SetStartValue(float value);
+		void SetEndValue(float value);
 		//开始动画
 		void Start(int durationMs, int fps = 90);
 		//动画是否已经停止
